XML Reading/Writing SCRAPPED - Cookies Implemented
So, after much research and deliberation, XML has been scrapped
completely. There are still some vestiges of XML code hanging around,
but they'll be removed as they are deemed unneeded.

All three web branches (Frontend, Backend, and Landing page) have been
merged into one folder so everything plays nicely (basically, cookies
need to function in the same path), plus it's easier to work with the
three of them if they're in the same folder.

So, basically, instead of storing everything in a file somewhere, we're
just making a cookie with an expiration date of a year and doing
everything on the fly.

We still need it to run on a webserver, though. So, good call Jordan.