Weather & RSS: Refresh after 5 minutes
We are now refreshing weather and RSS information every five minutes.
Also, it has been confirmed that once the weather code changes
(say...from thunderstorms to sunny), the new video will load and play.
However, if it stays the same, it will not replay the same video over
and over. Yay!